refactor(tests): extract k8s client helpers into dedicated k8sutil package
Move ConfigureRetryableTransport and CreateK8sClientWithConfig from
vmhelpers into tests/k8sutil/ so tests/common.go no longer imports the
VM-specific vmhelpers package. The vmhelpers package retains thin
wrappers for backward compatibility.

Partially generated by AI.
